==> jupyter_notebooks/09_check_correlated_feats.ipynb <==

"""
This notebook just checks there won't be any correlated features
with the label in the processed data
"""

%cd /afs/crc.nd.edu/user/m/msaebi/Public/chemistry/yield_rxn

import argparse
import logging
import os
import json
import math
import warnings
import pickle

import pandas as pd
import numpy as np
from collections import defaultdict

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import ShuffleSplit

import rdkit
from rdkit.ML.Descriptors import MoleculeDescriptors
import rdkit.Chem as Chem
from rdkit.Chem import Descriptors

warnings.filterwarnings("ignore")

yield_dict=defaultdict()
data_type='az'
with open('data/'+data_type+'/raw/'+data_type+'_reactions_data.json') as datafile:
    lines = json.load(datafile)

    for line in lines:
        if data_type=='az':
            name= reaction_num = line["reaction_Num"]
            r_yield=line['yield']['yield']
        
        elif data_type in ['dy','su']:
            name= reaction_num = line["Id"]
            r_yield=line['yield']
        
        yield_dict[int(name)]=r_yield


az_csv = pd.read_csv('data/'+data_type+'/old/'+data_type+'_data.csv')

for col in az_csv.columns:
    feat= az_csv[col].values
    y = az_csv['yield'].values
    corr= np.corrcoef(feat, y)
    #print(col, corr[0][1])
    if abs(corr[0][1])>0.5:
        print(col)
        print(corr)

old_data_dict= az_csv.transpose().to_dict(orient='list')

old_data_dict.keys()

az_old_data = open('data/'+data_type+'/old/'+data_type+'_reactions_data.txt','r').readlines()

feature_names= ['f_'+str(i) for i in range(len(az_old_data[0].strip('\n').split('\t')[1].split(',')))]
data_dict= defaultdict(lambda: defaultdict())
for line in az_old_data:
    Id, features = line.split('\t')
    features= features.split(',')
    #data_dict['id']= int(Id)
    #data_dict[int(Id)]['yield']= yield_dict[int(Id)]
    for i in range(len(feature_names)):
        data_dict[int(Id)][feature_names[i]]= float(features[i])


df =pd.DataFrame.from_dict(data_dict, orient='index')

for col in df.columns:
    feat= df[col].values
    y = az_csv['yield'].values
    corr= np.corrcoef(feat, y)
    #print(col, corr[0][1])
    if abs(corr[0][1])>0.5:
        print(col)
        print(corr)

len(feature_names)
